refactor: make user direct member of domain package
The subdivision into model seems unnecessary.
